Handle the modqueue more intelligently
Every item in the modqueue is now checked on each run through the loop
(still a max of 100 items are read). The only limit now is that an item
in the modqueue will only be acted on once - even if it is removed from
the queue and re-appears later (by being reported again, or caught by
the spam filter).